DIGG-58: Starting major refactor of Searchpage components
