--- components/Search/index.ts
@@ -1,6 +1,0 @@
-export * from "./EntryScape";
-export * from "./GlobalSearch";
-export * from "./SearchFilters";
-export * from "./SearchForm";
-export * from "./SearchProvider";
-export * from "./SearchResults";
