--- components/SearchFilter/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchFilter";
