+++ components/content/Search/SearchFilters/index.tsx
@@ -1,7 +1,9 @@
 import React, { useState } from "react";
-import { ESRdfType, ESType, SearchContextData, SearchFilter } from "..";
-import { CloseIcon as CloseIcon2, FilterIcon } from "../Icons";
+import { ESRdfType, ESType } from "@/utilities/entryScape";
+import { CloseIcon as CloseIcon2, FilterIcon } from "@/components/Icons";
 import useTranslation from "next-translate/useTranslation";
+import { SearchContextData } from "@/providers/SearchProvider";
+import { SearchFilter } from "@/components/content/Search/SearchFilters/SearchFilter";
 
 interface SearchFilterProps {
   showFilter: boolean;
