+++ components/index.ts
@@ -1,10 +1,5 @@
 export * from "./ApiExploring";
 export * from "./blocks";
 export * from "./CookieBanner";
-export * from "./EntrystoreProvider";
-export * from "./FileFormatBadge";
 export * from "./ReactPortal";
-export * from "./Search";
-export * from "./SearchFilter";
-export * from "./SearchHead";
 export * from "./StartPageComponents";
