+++ components/pages/ConceptPage.tsx
@@ -1,12 +1,12 @@
 import useTranslation from "next-translate/useTranslation";
 import React, { useContext, useEffect } from "react";
 import { SettingsContext } from "@/providers/SettingsProvider";
-import { EntrystoreContext } from "../EntrystoreProvider";
-import { initBreadcrumb } from "../../pages/_app";
+import { initBreadcrumb } from "@/pages/_app";
 import { useMatomo } from "@datapunt/matomo-tracker-react";
 import { useRouter } from "next/router";
 import Head from "next/head";
 import { hemvist } from "@/utilities";
+import { EntrystoreContext } from "@/providers/EntrystoreProvider";
 
 export const ConceptPage: React.FC<{ curi?: string; scheme?: string }> = ({
   curi,
