+++ pages/concepts/[concept]/[param].tsx
@@ -1,9 +1,9 @@
 import { useRouter } from "next/router";
 import { useContext } from "react";
-import { SettingsContext } from "../../../components";
-import EntrystoreProvider from "../../../components/EntrystoreProvider/EntrystoreProvider";
-import { ConceptPage } from "../../../components/pages/ConceptPage";
-import { useScript } from "../../../hooks/useScript";
+import { ConceptPage } from "@/components/pages/ConceptPage";
+import { useScript } from "@/hooks/useScript";
+import { SettingsContext } from "@/providers/SettingsProvider";
+import EntrystoreProvider from "@/providers/EntrystoreProvider";
 
 export default function Concept() {
   const { env } = useContext(SettingsContext);
