+++ pages/datasets/[dataSet]/[name]/apiexplore/[apieid].tsx
@@ -1,13 +1,10 @@
-import { DataSetExploreApiPage } from "../../../../../components/pages/DatasetExploreApiPage";
+import { DataSetExploreApiPage } from "@/components/pages/DatasetExploreApiPage";
 // export default DataSetExploreApiPage;
 import { useRouter } from "next/router";
 import { useContext } from "react";
-import { useScript } from "../../../../../hooks/useScript";
-import {
-  ApiIndexProvider,
-  EntrystoreProvider,
-  SettingsContext,
-} from "../../../../../components";
+import { useScript } from "@/hooks/useScript";
+import { ApiIndexProvider, EntrystoreProvider } from "@/components";
+import { SettingsContext } from "@/providers/SettingsProvider";
 
 export default function ExploreApiPage() {
   const { env } = useContext(SettingsContext);
