+++ pages/datasets/[dataSet]/index.tsx
@@ -1,12 +1,10 @@
 import { useRouter } from "next/router";
 import { useContext } from "react";
-import {
-  ApiIndexProvider,
-  EntrystoreProvider,
-  SettingsContext,
-} from "../../../components";
-import { DataSetPage } from "../../../components/pages/DataSetPage";
-import { useScript } from "../../../hooks/useScript";
+import { ApiIndexProvider } from "@/components";
+import { DataSetPage } from "@/components/pages/DataSetPage";
+import { useScript } from "@/hooks/useScript";
+import { SettingsContext } from "@/providers/SettingsProvider";
+import EntrystoreProvider from "@/providers/EntrystoreProvider";
 
 export default function DataSet() {
   const { env } = useContext(SettingsContext);
