+++ pages/specifications/index.ts
@@ -1,3 +1,3 @@
-import { SearchSpecificationsPage } from "../../components/pages/SearchSpecificationsPage";
+import { SearchSpecificationsPage } from "@/components/content/Search//SearchPage/SearchSpecificationsPage";
 
 export default SearchSpecificationsPage;
